backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.30.0
python-multipart>=0.0.6
assemblyai>=0.35.0
supabase>=2.10.0
pydantic>=2.5.3
python-dotenv>=1.0.0
aiofiles>=23.2.1
httpx>=0.27.0
websockets>=13.0
